+++ src/main/resources/messages.properties
@@ -208,7 +208,7 @@ tutoria.listar.resumen=Resumen de dudas
 tutoria.listar.confirmada=Confirmada
 
 tutoria.confirmar.confirmada=Tutoría confirmada
-tutoria.confirmar.confirmada.descripcion=Su tutoría
+tutoria.confirmar.confirmada.descripcion=Su tutoría para la asignatura <b>{0}</b> con <b>{1}</b>, ha sido confirmada.
 tutoria.confirmar.no.confirmada=Tutoría no confirmada
 tutoria.confirmar.no.confirmada.descripcion=Ha habido un error al confirmar la tutoría.
