+++ src/main/resources/tutorias-init.sql
@@ -1,1 +1,2 @@
-INSERT INTO TUTORIA (ID, resumenDudas, EMISOR_ID, DESTINATARIO_ID, confirmada, asignatura) VALUES (1, 'No tengo ni idea', 0,2, true, 'MTP');
+INSERT INTO TUTORIA (ID, resumenDudas, EMISOR_ID, DESTINATARIO_ID, confirmada, asignatura) VALUES (0, 'No tengo ni idea', 0,2, false, 'MTP');
+INSERT INTO TUTORIA (ID, resumenDudas, EMISOR_ID, DESTINATARIO_ID, confirmada, asignatura) VALUES (1, 'Dudas sobre los puntos 3, 4 y 6 ', 2,1, true, 'Matemática Discreta');
