+++ src/main/webapp/WEB-INF/jsp/portal/temporal.jsp
@@ -1,5 +1,8 @@
-<meta http-equiv="refresh" content="2; URL=${urlRedireccion}">
+<meta http-equiv="refresh" content="3; URL=${urlRedireccion}">
 
 <%@ taglib prefix="spring" uri="http://www.springframework.org/tags"%>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
 
+<%@ taglib prefix="tiles" uri="http://tiles.apache.org/tags-tiles"%>
+
+<h3>${texto}</h3>
