+++ src/main/webapp/WEB-INF/spring/webcontext/portal-context.xml
@@ -29,7 +29,7 @@
 	<context:component-scan base-package="es.ucm.fdi.espacios.business.boundary" />
 	<context:component-scan base-package="es.ucm.fdi.users.business.boundary" />
 	<context:component-scan base-package="es.ucm.fdi.acortador.business.boundary" />
-	<context:component-scan base-package="es.ucm.fdi.acortador.tutorias.boundary" />	
+	<context:component-scan base-package="es.ucm.fdi.tutorias.business.boundary" />	
 
 
 	<!-- Packages to search for @Controller beans definitions -->
@@ -79,7 +79,7 @@
 	<util:properties id="storagePrefs">
 		<prop key="rootPath">WEB-INF/storage</prop>
 <!-- 		<prop key="serviceURL">http://localhost:8080/portal/storage/</prop> -->
-		<prop key="serviceURL">http://localhost:8088/anuncios/storage/</prop>
+		<prop key="serviceURL">http://localhost:8088/portal/storage/</prop>
 	</util:properties>
