+++ src/main/webapp/WEB-INF/tiles/portal/definitions/tile-definition.xml
@@ -151,7 +151,7 @@
  		TUTORIAS
  	***********************************************************************-->
 	
-	<!-- Listado de reservas -->
+	<!-- Listado de tutorias -->
 	<definition name="listarTutorias" extends="baseLayout">
 		<put-attribute name="title" value="tutoria.listar.title" />
 		<put-attribute name="resources" value="/WEB-INF/jsp/portal/listarTutorias_resources.jsp" />
@@ -161,7 +161,7 @@
 	</definition>
 	
 	
-	<!-- Creador de reservas -->
+	<!-- Creador de tutorias -->
 	<definition name="editorTutorias" extends="baseLayoutWithoutJumbotron">
 		<put-attribute name="title" value="" />
 		<put-attribute name="resources" value="/WEB-INF/jsp/portal/editorTutorias_resources.jsp" />
@@ -169,15 +169,6 @@
 		<put-attribute name="tagline" value="" />
 		<put-attribute name="content" value="/WEB-INF/jsp/portal/editorTutorias.jsp" />
 	</definition>
-	
-	<!-- Calendario de reservas -->
-<!-- 	<definition name="calendarioReservas" extends="baseLayoutWithoutJumbotron"> -->
-<!-- 		<put-attribute name="title" value="listado.reservas.titulo" /> -->
-<!-- 		<put-attribute name="resources" value="/WEB-INF/jsp/portal/calendarioReservas_resources.jsp" /> -->
-<!-- 		<put-attribute name="heading" value="" /> -->
-<!-- 		<put-attribute name="tagline" value="" /> -->
-<!-- 		<put-attribute name="content" value="/WEB-INF/jsp/portal/calendarioReservas.jsp" /> -->
-<!-- 	</definition> -->
 	
 	
 	<!--***********************************************************************
@@ -207,7 +198,6 @@
 	<definition name="temporal" extends="baseLayout">
 		<put-attribute name="title" value="temporal.title" />
 		<put-attribute name="heading" expression="${texto1}	" />
-		<put-attribute name="tagline" expression="${texto2}" />
 		<put-attribute name="content" value="/WEB-INF/jsp/portal/temporal.jsp" />
 	</definition>
