+++ src/main/webapp/static/404.html
@@ -49,7 +49,7 @@
     <div class="dialog">
       <h1>Ups.</h1>
       <h2>Algo se nos ha perdido por el camino!</h2>
-      <h3>Mejor dirígete a la <a href="/anuncios/">página principal</a>.</h3>
+      <h3>Mejor dirígete a la <a href="/portal/">página principal</a>.</h3>
     </div>
   
 </body></html>
