+++ src/test/resources/testSolapamientoReservas.xml
@@ -13,7 +13,7 @@
 </thead><tbody>
 <tr>
 	<td>open</td>
-	<td>/anuncios/reservas/nuevo</td>
+	<td>/portal/reservas/nuevo</td>
 	<td></td>
 </tr>
 <tr>
